Interim commit of major system refactor
Still a lot of work to be done, but major progress made on better-
aligning original implementation with Foundry conventions, as well as
separation of models, views, and controllers.

A great deal of work left to do, and this commit likely contains
breaking changes with pieces that are not included in it (e.g. non-
character actors), so should not be considered functional on its own.